requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==23.2.0
Authlib==1.3.1
blinker==1.8.2
blis==1.0.1
boto3==1.35.42
botocore==1.35.42
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==2.2.1
cohere==5.11.0
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
databricks-sdk==0.34.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
docker==7.1.0
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
et-xmlfile==1.1.0
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.35.0
google-pasta==0.2.0
graphene==3.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-health-checking==1.62.3
grpcio-tools==1.62.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.25.2
idna==3.10
importlib-metadata==6.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.3
langchain-cohere==0.3.1
langchain-community==0.3.2
langchain-core==0.3.11
langchain-experimental==0.3.2
langchain-text-splitters==0.3.0
langcodes==3.4.1
langsmith==0.1.135
language_data==1.2.0
lxml==5.3.0
Mako==1.3.5
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.17.0
mlflow-skinny==2.17.0
mock==4.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.9.3
networkx==3.4.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==0.28.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pathos==0.3.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.3.6
pox==0.3.5
ppft==1.7.6.9
preshed==3.0.9
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.11
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
srsly==2.4.8
streamlit==1.39.0
streamlit-chat==0.1.1
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
triton==3.1.0
typer==0.12.5
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
wasabi==1.1.3
watchdog==5.0.3
weasel==0.4.1
weaviate==0.1.0
weaviate-client==4.8.1
Werkzeug==3.0.4
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.15.4
zipp==3.20.2
